refactor: mount dj module at /dj routes
