--- src/app/login/page.tsx
@@ -1,1 +1,0 @@
-export { default } from '@dj/app/login/page';
