--- src/app/page.tsx
@@ -1,1 +1,0 @@
-export { default } from '@dj/app/queue/page';
