+++ src/middleware.ts
@@ -3,7 +3,7 @@ import { hasAdminCookie } from '@core/api/auth';
 
 export function middleware(req: NextRequest) {
   const { pathname } = req.nextUrl;
-  const isAdminRoute = pathname.startsWith('/admin');
+  const isAdminRoute = pathname.startsWith('/dj/admin');
   const isRequestsRoute = pathname.startsWith('/api/requests');
   const isPublicRequest =
     isRequestsRoute && pathname === '/api/requests' && ['GET', 'POST'].includes(req.method);
@@ -17,7 +17,7 @@ export function middleware(req: NextRequest) {
       return NextResponse.next();
     }
     if (isAdminRoute) {
-      return NextResponse.redirect(new URL('/login', req.url));
+      return NextResponse.redirect(new URL('/dj/login', req.url));
     }
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
@@ -26,5 +26,5 @@ export function middleware(req: NextRequest) {
 }
 
 export const config = {
-  matcher: ['/admin', '/admin/:path*', '/api/requests/:path*'],
+  matcher: ['/dj/admin', '/dj/admin/:path*', '/api/requests/:path*'],
 };
